Generic Object Pooling
Implement Generic Object Pooling for those derived from Game Object.
Must have a base constructor taking in Scene* pointer.

Game Object active state get and set. Used to determine if an object should be updated, rendered and have inputs registered.

Projectiles going out of bounds are now set to inactive, and re-activated when needed. Seems to improve performance with stupid amounts of projectiles, but there appears to be some catches to this. If enough collisions occur on the same frame it will not find any de-activated objects and thus add another to the pool. Will add some removal process later based on % of pool active perhaps.